+++ NEWS.md
@@ -1,3 +1,14 @@
+
+# emisc 0.0.3
+
+## New Functions
+
+Added the `write_latex` and `latex_markup` functions.
+
+`write_latex` saves all or selected dataframe columns to a LaTeX file in rows, rather than as a table.
+
+`latex_markup` adds LaTeX markup (e.g. section tags) to selected columns to assist with printing.
+
 # emisc 0.0.2
 
 ## Bug Fixes